Calling function by function name stored in the database
