+++ .gitignore
@@ -11,3 +11,4 @@
 /C_Sharp_Server/C_Sharp_Server/auth.json
 /C_Sharp_Server/Lynxmotion
 /C_Sharp_Server/TestLynxmotion
+/C_Sharp_Server/TestFunction
